Move create_tombstone_for_form to models.py
