+++ corehq/apps/tombstones/models.py
@@ -1,7 +1,20 @@
+from datetime import UTC, datetime
+
 from django.db import models
 
 from corehq.sql_db.models import PartitionedModel, RequireDBManager
 from corehq.sql_db.util import split_list_by_db_partition
+
+
+def create_tombstone_for_form(form):
+    from corehq.form_processor.models.forms import XFormInstance
+
+    return Tombstone(
+        doc_id=form.form_id,
+        object_class_path=f'{XFormInstance.__module__}.{XFormInstance.__qualname__}',
+        domain=form.domain,
+        deleted_on=form.deleted_on or datetime.now(tz=UTC),
+    )
 
 
 class TombstoneObjectManager(RequireDBManager):
